move interpolate into utils
